+++ CHANGELOG.md
@@ -2,8 +2,21 @@
 
 _Note: Only from BubbleOS v0.2.2 will changes be recorded._
 
-## v0.2.3
+Please note that not all versions will have respectible executables with them.
+
+## Build 23 (v0.2.3)
+
+### Added
 
   - Extracted argument repition into seperate function `_singleParam`.
+  - Added `CHANGELOG.md`
+
+### Changed
+
+  - Updated the `error` function to make it more reuseable and lightweight
   - Updated `cd` command to handle no directories
   - Improved code reusability
+
+### Removed
+
+  - Deprecated the `RECOGNIZED_COMMANDS` constant.